requirements.txt
absl-py==0.6.1
acme==0.26.0
africastalking==1.1.5
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
authy==2.2.3
backcall==0.1.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bleach==3.0.2
blinker==1.4
Brlapi==0.6.6
bs4==0.0.1
cachetools==3.1.0
certbot==0.26.1
certbot-nginx==0.25.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
command-not-found==0.3
config==0.4.0
ConfigArgParse==0.11.0
configobj==5.0.6
contextlib2==0.5.5
crypto==1.4.1
cryptography==2.6
cupshelpers==1.0
cycler==0.10.0
Cython==0.29
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
entrypoints==0.2.3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
future==0.15.2
gast==0.2.0
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-bigquery==1.10.0
google-cloud-core==0.29.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.16.1
h5py==2.9.0
httplib2==0.9.2
idna==2.8
importlib==1.0.4
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
josepy==1.1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
mock==2.0.0
mysqlclient==1.4.1
Naked==0.1.31
nbconvert==5.4.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
nltk==3.4
notebook==5.7.1
numpy==1.16.1
oauth==1.0.1
olefile==0.45.1
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parsedatetime==2.4
parso==0.3.1
passlib==1.7.1
pathlib==1.0.1
pbr==3.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.7.0
ptyprocess==0.6.0
pusher==2.1.3
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
PyICU==1.9.8
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.3.0
pyRFC3339==1.0
python-apt==1.6.2
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2018.9
pyxdg==0.25
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
reportlab==3.4.0
requests==2.21.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
rsa==4.0
scapy==2.4.0
scapy-python3==0.26
schema==0.7.0
scikit-learn==0.20.0
scipy==1.1.0
SecretStorage==2.3.1
Send2Trash==1.5.0
shellescape==3.4.1
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.7.3
SQLAlchemy==1.3.1
system-service==0.3
systemd-python==234
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
Whoosh==2.7.4
widgetsnbextension==3.4.2
WTForms==2.2.1
xkit==0.0.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
